// src/app/model/params-file.ts
export interface ParamsFile {
  user_defined_cst_file?: string;
  res_chain?: string;
  cat_ID?: string;
  id?: string;
  ligand_params_file?: string;
  design_ligand_name_out?: string;
  other_ligands?: string;
  design_cst?: null;
  res_ligand_chain?: string;
  protein_renumber_file?: string;
  constrain_info?: string;
  inputs_pos_file?: string;
  res_ligand_name?: string;
  ligand_file?: string;
  analysis_param?: string;
  match_cst_file?: string;
  job_name?: string;
  user_email?: string;
  UM_pdb_count?: number;
  filtered_designed_ligandname_out?: string;
}